+++ python/python/glide/async_commands/transaction.py
@@ -1067,6 +1067,23 @@ class BaseTransaction:
         """
         return self.append_command(RequestType.SInter, keys)
 
+    def sinterstore(
+        self: TTransaction, destination: str, keys: List[str]
+    ) -> TTransaction:
+        """
+        Stores the members of the intersection of all given sets specified by `keys` into a new set at `destination`.
+
+        See https://valkey.io/docs/latest/commands/sinterstore for more details.
+
+        Args:
+            destination (str): The key of the destination set.
+            keys (List[str]): The keys from which to retrieve the set members.
+
+        Command response:
+            int: The number of elements in the resulting set.
+        """
+        return self.append_command(RequestType.SInterStore, [destination] + keys)
+
     def sdiff(self: TTransaction, keys: List[str]) -> TTransaction:
         """
         Computes the difference between the first set and all the successive sets in `keys`.